Improve performance by using useQuery to fetch the pokemon data/cache
